feat(web): renderer switch (Text/Image) + image thumbnail + lightbox
